Fix sleep large number

+++ ext-src/swoole_runtime.cc
@@ -1611,9 +1611,9 @@ static PHP_FUNCTION(swoole_sleep) {
     }
 
     if (Coroutine::get_current()) {
-        RETURN_LONG(System::sleep((double) num) < 0 ? num : 0);
+        RETURN_LONG(System::sleep((double) (unsigned int) num) < 0 ? num : 0);
     } else {
-        RETURN_LONG(php_sleep(num));
+        RETURN_LONG(php_sleep((unsigned int) num));
     }
 }
